+++ spec/api_paths_spec.rb
@@ -1,7 +1,6 @@
 require 'spec_helper'
 
 describe 'simple api with prefix' do
-
   before :all do
     class ApiWithPrefix < Grape::API
       prefix :api
@@ -16,7 +15,6 @@ describe 'simple api with prefix' do
       mount ApiWithPrefix
       add_swagger_documentation
     end
-
   end
 
   def app
@@ -63,5 +61,4 @@ describe 'simple api with prefix' do
       )
     end
   end
-
 end