+++ spec/markdown/kramdown_adapter_spec.rb
@@ -1,7 +1,6 @@
 require 'spec_helper'
 
 describe GrapeSwagger::Markdown::KramdownAdapter do
-
   context 'initialization' do
     it 'uses GFM as default input and disable coderay' do
       adapter = GrapeSwagger::Markdown::KramdownAdapter.new