+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Next Release
 
 * [#196](https://github.com/tim-vandecasteele/grape-swagger/pull/196): If `:type` is omitted, see if it's available in `:using` - [@jhollinger](https://github.com/jhollinger).
+* [#200](https://github.com/tim-vandecasteele/grape-swagger/pull/200): Treat `type: Symbol` as string form parameter - [@ypresto](https://github.com/ypresto).
 * Your contribution here.
 
 ### 0.9.0 (December 19, 2014)